+++ aikau-sandpit-application/src/main/webapp/WEB-INF/classes/alfresco/site-webscripts/imports/sandpit.lib.js
@@ -143,7 +143,7 @@ function buildPageModel(data) {
                                  name: "alfresco/header/Title",
                                  align: "left",
                                  config: {
-                                    label: "Aikau Sandpit - " + data.title,
+                                    label: "Aikau Sandpit - " + msg.get(data.title),
                                     setBrowserTitle: "Aikau Sandpit"
                                  }
                               },