+++ aikau/src/main/resources/alfresco/services/SiteService.js
@@ -60,7 +60,8 @@ define(["dojo/_base/declare",
        *
        * @instance
        * @since 1.0.32
-       * @listens module:alfresco/core/topics~event:DELETE_SITE
+       * @listens module:alfresco/core/topics#DELETE_SITE
+       * @listens module:alfresco/core/topics#BECOME_SITE_MANAGER
        */
       registerSubscriptions: function alfresco_services_SiteService__registerSubscriptions() {
          this.alfSubscribe("ALF_GET_SITES", lang.hitch(this, this.getSites));
@@ -68,7 +69,7 @@ define(["dojo/_base/declare",
          this.alfSubscribe("ALF_GET_SITE_MEMBERSHIPS", lang.hitch(this, this.getSiteMemberships));
          this.alfSubscribe("ALF_GET_SITE_DETAILS", lang.hitch(this, this.getSiteDetails));
          this.alfSubscribe("ALF_UPDATE_SITE_DETAILS", lang.hitch(this, this.updateSite));
-         this.alfSubscribe("ALF_BECOME_SITE_MANAGER", lang.hitch(this, this.becomeSiteManager));
+         this.alfSubscribe(topics.BECOME_SITE_MANAGER, lang.hitch(this, this.becomeSiteManager));
          this.alfSubscribe("ALF_JOIN_SITE", lang.hitch(this, this.joinSite));
          this.alfSubscribe("ALF_REQUEST_SITE_MEMBERSHIP", lang.hitch(this, this.requestSiteMembership));
          this.alfSubscribe("ALF_LEAVE_SITE", lang.hitch(this, this.leaveSiteRequest));
@@ -287,8 +288,8 @@ define(["dojo/_base/declare",
        *
        * @instance
        * @param {object} payload
-       * @fires module:alfresco/core/topics~event:NAVIGATE_TO_PAGE
-       * @fires module:alfresco/core/topics~event:RELOAD_DATA_TOPIC
+       * @fires module:alfresco/core/topics#NAVIGATE_TO_PAGE
+       * @fires module:alfresco/core/topics#RELOAD_DATA_TOPIC
        */
       onActionDeleteSiteSuccess: function alfresco_services_SiteService__onActionDeleteSiteSuccess(payload) {
          var subscriptionHandle = lang.getObject("requestConfig.subscriptionHandle", false, payload);
@@ -444,7 +445,7 @@ define(["dojo/_base/declare",
             this.serviceXhr({url : url,
                              data: data,
                              method: "POST",
-                             successCallback: this.reloadData,
+                             successCallback: config.reloadPage ? this.reloadPage : this.reloadData,
                              callbackScope: this});
          }
          else
@@ -735,19 +736,25 @@ define(["dojo/_base/declare",
        * This is a catch all success handler for both the join site and become site manager. It simply reloads
        * the current page. It is ** INCORRECTLY ** assumed that the current user will always be on the site
        * referenced in the request. This method needs to be updated accordingly.
+       *
+       * @instance
+       * @fires module:alfresco/core/topics#RELOAD_PAGE
        */
       reloadPage: function alfresco_services_SiteService__reloadPage(response, requestConfig) {
          /*jshint unused:false*/
          // TODO: Check user in request is current user and that site in request is current site
-         this.alfPublish("ALF_RELOAD_PAGE", {});
+         this.alfPublish(topics.RELOAD_PAGE);
       },
 
       /**
        * This is a method that reloads DocList data
+       *
+       * @instance
+       * @fires module:alfresco/core/topics#RELOAD_DATA_TOPIC
        */
       reloadData: function alfresco_services_SiteService__reloadData(response, requestConfig) {
          /*jshint unused:false*/
-         this.alfPublish("ALF_DOCLIST_RELOAD_DATA", {});
+         this.alfPublish(topics.RELOAD_DATA_TOPIC);
       },
 
       /**