+++ aikau/src/test/resources/alfresco/services/actions/DownloadAsZipTest.js
@@ -111,10 +111,7 @@ registerSuite(function(){
       },
 
       "Test non-legacy action version": function() {
-         return browser.findByCssSelector("#ACTIONS_ITEM_1_MENU_text")
-            .click()
-         .end()
-         .findAllByCssSelector("#ACTIONS_ITEM_1_DOWNLOAD_AS_ZIP")
+         return browser.findAllByCssSelector("#ACTIONS_ITEM_1_DOWNLOAD_AS_ZIP")
             .click()
          .end()
          .findAllByCssSelector("#ARCHIVING_DIALOG.dialogDisplayed")